+++ src/main/java/com/mongodb/shardsync/command/drop/DropDatabasesCommand.java
@@ -1,23 +1,18 @@
 package com.mongodb.shardsync.command.drop;
 
+import java.util.concurrent.Callable;
+
 import com.mongodb.shardsync.ShardConfigSync;
 import com.mongodb.shardsync.SyncConfiguration;
-import com.mongodb.shardsync.command.DropCommand;
+
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
-import picocli.CommandLine.Option;
-
-import java.util.concurrent.Callable;
 
 @Command(name = "databases", description = "Drop databases on destination cluster")
 public class DropDatabasesCommand implements Callable<Integer> {
     
     @CommandLine.ParentCommand
     private DropCommand parent;
-    
-    @Option(names = {"--include-config"}, 
-            description = "Also drop config metadata")
-    private boolean includeConfig;
     
     @Override
     public Integer call() throws Exception {
@@ -26,11 +21,7 @@ public class DropDatabasesCommand implements Callable<Integer> {
         ShardConfigSync sync = new ShardConfigSync(config);
         sync.initialize();
         
-        if (includeConfig) {
-            sync.dropDestinationDatabasesAndConfigMetadata();
-        } else {
-            sync.dropDestinationDatabases();
-        }
+        sync.dropDestinationDatabasesAndConfigMetadata();
         
         return 0;
     }
