+++ src/main/java/com/mongodb/shardsync/command/sync/SyncCommand.java
@@ -1,8 +1,9 @@
-package com.mongodb.shardsync.command;
+package com.mongodb.shardsync.command.sync;
 
 import com.mongodb.shardsync.ShardConfigSyncApp;
 import com.mongodb.shardsync.SyncConfiguration;
-import com.mongodb.shardsync.command.sync.*;
+import com.mongodb.shardsync.command.AtlasOptionsMixin;
+
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
